Move CheckInstallationSourceImageTask to payload
This task is live_image specific task so move it under live_image
payload folder.
