+++ pyanaconda/modules/payload/payloads/live_image/live_image.py
@@ -32,9 +32,10 @@ from pyanaconda.modules.payload.base.handler_base import PayloadHandlerBase
 from pyanaconda.modules.payload.base.initialization import CopyDriverDisksFilesTask
 from pyanaconda.modules.payload.payloads.live_image.live_image_interface import \
     LiveImageHandlerInterface
-from pyanaconda.modules.payload.live.initialization import CheckInstallationSourceImageTask, \
-    SetupInstallationSourceImageTask, UpdateBLSConfigurationTask, \
-    TeardownInstallationSourceImageTask
+from pyanaconda.modules.payload.payloads.live_image.initialization import \
+    CheckInstallationSourceImageTask
+from pyanaconda.modules.payload.live.initialization import SetupInstallationSourceImageTask, \
+    UpdateBLSConfigurationTask, TeardownInstallationSourceImageTask
 from pyanaconda.modules.payload.live.utils import get_kernel_version_list, \
     get_kernel_version_list_from_tar, url_target_is_tarfile
 from pyanaconda.modules.payload.live.installation import InstallFromImageTask, InstallFromTarTask
